Reading .12d
Function that reads the .12d file and stores the values ​​in a matrix where each column is a matrix of QTFs for a given wave incidence. For example, if the .12d file contains wave incidences 0deg and 30deg, the first column of the matrix will be the QTF matrix for 0deg, and the second for 30deg.